01_Applications/04_tubeauto/requirements.txt
streamlit>=1.28.0
openai>=1.12.0
opencv-python>=4.8.0
pydub>=0.25.1
python-dotenv>=1.0.0
numpy>=2.0.0,<2.3.0
pillow>=10.0.0
requests>=2.31.0

